Fix problem with trying to drop tables when they don't exist

+++ setup.rb
@@ -23,8 +23,8 @@ puts "CONNECTED"
 puts "Setting up Database (recreating tables) ..."
 
 ActiveRecord::Schema.define do
-  drop_table :stores
-  drop_table :employees
+  drop_table :stores if ActiveRecord::Base.connection.table_exists?(:stores)
+  drop_table :employees if ActiveRecord::Base.connection.table_exists?(:employees)
   create_table :stores do |t|
     t.column :name, :string
     t.column :annual_revenue, :integer
